Make z_once() local to avoid conditional external symbols.
